# openapi.yml
openapi: 3.1.0
info:
  title: Task API
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: Get paginated list of tasks
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 0
          description: Number of tasks per page (minimum 0)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example:
                total: 1
                totalPages: 1
                page: 1
                pageSize: 200
                tasks: [{ title: "New Task", description: "Description for the new task", completed: false }]

    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "New Task"
              description: "Description for the new task"
              completed: false
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              example:
                id: 1
                title: "New Task"
                description: "Description for the new task"
                completed: false

  /tasks/{id}:
    get:
      summary: Get details of a specific task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example:
                id: 1
                title: "New Task"
                description: "Description for the new task"
                completed: false
        '404':
          description: Task not found

    put:
      summary: Update an existing task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "Updated Title"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              example:
                id: 1
                title: "Updated Title"
                description: "Description for the new task"
                completed: false
        '404':
          description: Task not found
        '400':
          description: Bad Request
          content:
            application/json:
              example:
                error: "Invalid Task: title required"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              example:
                error: "Unable to modify Task id"

    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '404':
          description: Task not found